refactor(hybrid/h3.5): migrar useCachedQuery e cachedServices para nova engine + remover queryCache.js legado
- useCachedQuery: troca import de @shared/utils/queryCache por webQueryCache (nova engine)
- cachedServices: troca import de @shared/utils/queryCache por webQueryCache
- useDashboardContext: troca import de @shared/utils/queryCache por @shared/hooks/useCachedQuery
- src/shared/utils/queryCache.js: REMOVIDO (substituido por packages/shared-data + src/shared/platform/)

Interface publica de useCachedQuery (CON-006) permanece identica.
Validacao: lint OK, 543/543 testes passam, build OK.
